Fix destoyed Redux Modules
